requirements.txt
mitmproxy~=9.0.1
motor~=3.1.2
python-dotenv~=1.0.0
pymongo~=4.3.3
python-magic
python-magic-bin==0.4.14
zstandard==0.21.0
